more doubles in place of floats

+++ valhalla/meili/match_result.h
@@ -17,11 +17,11 @@ struct MatchResult {
   // Coordinate of the match point
   midgard::PointLL lnglat;
   // Distance from measurement to the match point
-  float distance_from;
+  double distance_from;
   // Which edge this match point stays
   baldr::GraphId edgeid;
   // Percentage distance along the edge
-  float distance_along;
+  double distance_along;
   // Optional epoch time for this match point copied from original measurement
   double epoch_time;
   // Sequential state id
